Actualiza entidad User

+++ src/main/java/com/carlosarroyoam/userservice/dto/CreateUserRequest.java
@@ -22,7 +22,7 @@ public class CreateUserRequest implements Serializable {
 	@NotNull(message = "Age should not be null")
 	@Min(message = "Age should be min 18", value = 18)
 	@Max(message = "Age should be max 100", value = 100)
-	private Integer age;
+	private Byte age;
 
 	@NotBlank(message = "Email should not be blank")
 	@Email(message = "Email should be an valid email address")
